requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6